Move metabox functionalities out of MasterSite.php
This is part of our efforts to make this file shorter
