+++ src/MasterSite.php
@@ -123,8 +123,6 @@ class MasterSite extends TimberSite
         add_action('wp_enqueue_scripts', [PublicAssets::class, 'enqueue_js']);
         add_filter('safe_style_css', [$this, 'set_custom_allowed_css_properties']);
         add_filter('wp_kses_allowed_html', [$this, 'set_custom_allowed_attributes_filter'], 10, 2);
-        add_action('add_meta_boxes', [$this, 'add_meta_box_search'], 10, 2);
-        add_action('save_post', [$this, 'save_meta_box_search'], 10, 2);
         add_action('save_post', [$this, 'set_featured_image'], 10, 2);
         add_filter('wp_insert_post_data', [$this, 'require_post_title'], 10, 1);
         // Save "p4_global_project_tracking_id" on post save.
@@ -1098,54 +1096,6 @@ class MasterSite extends TimberSite
     }
 
     /**
-     * Creates a Metabox on the side of the Add/Edit Post/Page
-     * that is used for applying weight to the current Post/Page in search results.
-     *
-     * @param string  $post_type Post type.
-     * @param WP_Post|WP_Comment $post      Post object.
-     * phpcs:disable SlevomatCodingStandard.Functions.UnusedParameter -- add_meta_boxes callback
-     */
-    public function add_meta_box_search(string $post_type, $post): void
-    {
-        add_meta_box(
-            'meta-box-search',
-            'Search',
-            [$this, 'view_meta_box_search'],
-            ['post', 'page'],
-            'side',
-            'default',
-            [$post]
-        );
-    }
-    // phpcs:enable SlevomatCodingStandard.Functions.UnusedParameter
-
-    /**
-     * Renders a Metabox on the side of the Add/Edit Post/Page.
-     *
-     * @param WP_Post $post The currently Added/Edited post.
-     * phpcs:disable Generic.WhiteSpace.ScopeIndent
-     */
-    public function view_meta_box_search(WP_Post $post): void
-    {
-        $weight = get_post_meta($post->ID, 'weight', true);
-        $options = get_option('planet4_options');
-
-        echo '<label for="my_meta_box_text">'
-            . esc_html__('Weight', 'planet4-master-theme-backend')
-            . ' (1-' . esc_attr(Search\Search::DEFAULT_MAX_WEIGHT) . ')</label>
-                <input id="weight" type="text" name="weight" value="' . esc_attr($weight) . '" />';
-
-        $script_data = [
-            'act_page' => $options['act_page'] ?? null,
-            'action_weight' => Search\Search::DEFAULT_ACTION_WEIGHT,
-            'page_weight' => Search\Search::DEFAULT_PAGE_WEIGHT,
-        ];
-
-        do_action('enqueue_metabox_search_script', $script_data);
-    }
-    // phpcs:enable Generic.WhiteSpace.ScopeIndent
-
-    /**
      * Applies filters for list of users in dropdown
      *
      * @param Array|null $args The filter options and values.
@@ -1225,48 +1175,6 @@ class MasterSite extends TimberSite
         }
 
         return $user;
-    }
-
-    /**
-     * Saves the Search weight of the Post/Page.
-     *
-     * @param int     $post_id The ID of the current Post.
-     * @param WP_Post $post The current Post.
-     */
-    public function save_meta_box_search(int $post_id, WP_Post $post): void
-    {
-        global $pagenow;
-
-        // Ignore autosave.
-        if (defined('DOING_AUTOSAVE') && DOING_AUTOSAVE) {
-            return;
-        }
-        // Check user's capabilities.
-        if (!current_user_can('edit_post', $post_id)) {
-            return;
-        }
-        // Make sure there's input.
-        $weight = filter_input(
-            INPUT_POST,
-            'weight',
-            FILTER_VALIDATE_INT,
-            [
-                'options' => [
-                    'min_range' => Search\Search::DEFAULT_MIN_WEIGHT,
-                    'max_range' => Search\Search::DEFAULT_MAX_WEIGHT,
-                ],
-            ]
-        );
-
-        // If this is a new Page then set default weight for it.
-        if (!$weight && 'post-new.php' === $pagenow) {
-            if ('page' === $post->post_type) {
-                $weight = Search\Search::DEFAULT_PAGE_WEIGHT;
-            }
-        }
-
-        // Store weight.
-        update_post_meta($post_id, 'weight', $weight);
     }
 
     /**
